requirements.txt
Django==3.1
Pillow==7.2.0
psycopg2-binary==2.8.5
celery==4.4.7

djangorestframework==3.12.1
markdown==3.1.1
django-filter==2.4.0
pyyaml>=5.1
uritemplate==3.0.1

gunicorn==19.9.0

redis==3.5.3

django-reversion==3.0.8

dj-database-url==0.5.0
python-json-logger==0.1.11

drf-nested-routers==0.91

django-cors-headers==3.5.0

djangorestframework-simplejwt==4.4.0

requests==2.24.0
pydantic==1.7.1
django-bulk-update-or-create==0.2.0

pytest==6.0.1
pytest-django==3.9.0
pytest-cov==2.10.0
pytest-mock==3.2.0
coverage==5.2.1
hypothesis==5.29.0
